pend : useSpendingForm 추상화 작업 중단 - 타입추론 방법을 체크
